Split Stores from the mock file
